refactor: refactoring record component
