+++ src/components/record/RecordCard.tsx
@@ -7,7 +7,6 @@ import DeleteModalContainer from '@/components/Common/DeleteModal';
 import EditBox from '@/components/Common/EditBox';
 import Modal from '@/components/Common/Modal';
 import RecordInfo from '@/components/Record/RecordInfo';
-import UpdateTemplate from '@/components/UpdateTemplate';
 import { MainContext } from '@/store';
 
 type RecordCardPropsType = {
@@ -51,7 +50,7 @@ export default function RecordCard({ title, id }: RecordCardPropsType) {
           setIsOpen={setRecordTemplateOpen}
           height={isSmallScreen ? 720 : undefined}
         >
-          <RecordInfo id={id} />
+          <RecordInfo id={id} isEditing={false} />
         </Modal>
       )}
       {editModalOpen && (
@@ -59,7 +58,7 @@ export default function RecordCard({ title, id }: RecordCardPropsType) {
           setIsOpen={setEditModalOpen}
           height={isSmallScreen ? 720 : undefined}
         >
-          <UpdateTemplate id={id} />
+          <RecordInfo id={id} isEditing={true} />
         </Modal>
       )}
       {deleteModalOpen && (
