+++ src/components/record/RecordContents.tsx
@@ -3,10 +3,9 @@ import { useContext } from 'react';
 
 import EmptyQuestion from '@/assets/EmptyQuestion.svg';
 import Loading from '@/components/Common/Loading';
+import UpdateQuestion from '@/components/UpdateTemplate/UpdateQuestion';
 import useRecordDetail from '@/hooks/useRecordDetail';
 import { MainContext } from '@/store';
-
-import UpdateQuestion from '../UpdateTemplate/UpdateQuestion';
 
 export default function RecordContents() {
   const { selectedRecordCardId } = useContext(MainContext);
