+++ src/components/record/RecordInfo.tsx
@@ -1,16 +1,30 @@
+import Loading from '@/components/Common/Loading';
 import RecordContents from '@/components/Record/RecordContents';
+import RecordModalHeader from '@/components/Record/RecordModalHeader';
+import UpdateTemplate from '@/components/UpdateTemplate';
 import UpdateTemplateTitle from '@/components/UpdateTemplate/UpdateTemplateTitle';
-
-import RecordModalHeader from './RecordModalTitle';
+import useRecordDetail from '@/hooks/useRecordDetail';
 
 type RecordInfo = {
   id: number;
+  isEditing: boolean;
 };
 
-export default function RecordInfo({ id }: RecordInfo) {
-  return (
+export default function RecordInfo({ id, isEditing }: RecordInfo) {
+  const { recordDetailData } = useRecordDetail(id);
+
+  if (!recordDetailData) {
+    return (
+      <>
+        <Loading />
+      </>
+    );
+  }
+  return isEditing ? (
+    <UpdateTemplate id={id} recordDetailData={recordDetailData} />
+  ) : (
     <>
-      <UpdateTemplateTitle />
+      <UpdateTemplateTitle id={id} isEditing={isEditing} />
       <RecordModalHeader id={id} />
       <RecordContents />
     </>
