+++ src/components/record/index.tsx
@@ -1,9 +1,9 @@
 import { useMediaQuery } from '@chakra-ui/react';
-import { useCallback, useContext, useEffect, useState } from 'react';
+import { useCallback, useContext, useState } from 'react';
 
 import Modal from '@/components/Common/Modal';
-import TypeSelector from '@/components/Record//TypeSelector';
 import RecordListContainer from '@/components/Record/RecordListContainer';
+import TypeSelector from '@/components/Record/TypeSelector';
 import Template from '@/components/Template';
 import { MainContext } from '@/store';
 
@@ -13,20 +13,8 @@ export default function Record() {
   const changeListType = useCallback((type: string) => {
     setTemplateCategory(type);
   }, []);
-  const {
-    recordModalOpen,
-    setRecordModalState,
-    selectedTemplateTitle,
-    setSelectedTemplateTitle,
-    setQuestionList,
-  } = useContext(MainContext);
-
-  useEffect(() => {
-    if (!recordModalOpen) {
-      setSelectedTemplateTitle('');
-      setQuestionList([]);
-    }
-  }, [recordModalOpen, setQuestionList, setSelectedTemplateTitle]);
+  const { recordModalOpen, setRecordModalState, selectedTemplateTitle } =
+    useContext(MainContext);
 
   return (
     <>
